+++ DataMaps/Properties/AssemblyInfo.cs
@@ -3,5 +3,5 @@ using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("Pathoschild.Stardew.DataMaps")]
 [assembly: Guid("cdeef8af-15cb-4955-8163-e446f212c2f4")]
-[assembly: AssemblyVersion("1.1.0")]
-[assembly: AssemblyFileVersion("1.1.0")]
+[assembly: AssemblyVersion("1.2.0")]
+[assembly: AssemblyFileVersion("1.2.0")]